Major refactoring
I think this is a major improvement, but I'm still not happy with it. Card selection/removal still seems janky, as does the whole UNIQUE_CARDS/uniqueCards business. The view (i.e., DOM) code is still scattered throughout and there are a lot of side effects and logic hidden beneath seemingly innocuous functions.